Get sync controller to download source tarball
This fills in enough of the sync controller for it to fetch a sync's
source repo, from the source controller. The next step is to apply the
downloaded tarball to the cluster; after that, keeping the status up
to date, and embellishments like keeping the tarball around if the
target revision hasn't changed.

For now I'm using the source-controller, so this needs the CRDs from

    https://github.com/fluxcd/source-controller/tree/master/config/crd

to be applied in the cluster, as well as the source controller
deployment and service from

    https://github.com/fluxcd/source-controller/tree/master/config/manager

(the image tag may need to be adjusted in the deployment)