Refer to a URL in Sync rather than a resource
This changes the design such that Syncs point at a URL rather than a
source, explicitly. That means you, or another controller, can point
it at the "latest" URL of a source, or a tarball hosted somewhere else
(like GitHub, though it provides ZIP files rather than tarballs).

This simplifies the mechanism -- no need to look up or watch Sources
-- while achieving the same effect, and being more flexible.

The Sync spec also lets you supply paths to sync within the package
filesystem, since much of the time the YAMLs are in nested
directories.