Give Sync status information
I want to be able to identify when a Sync has successfully applied a
particular revision. This necessitates two things:

 - a record of what was last attempted
 - a record of whether it worked

In general the URL may not (conveniently) identify a particular
revision (i.e., what was last attempted), so the _revision_ goes into
the spec, to be copied into the status.