Make Sync apply step contingent
Now that I update the status of a Sync after an apply, reconcile get
triggered in a loop (since updating the status results in the watch
firing and the resource being requeued).

To stop this happening, the apply step -- which ultimately updates the
status -- should be contingent on whether there's work to actually
do. This commit takes that to mean either the resource changed in
substance (the generation was bumped) or the sync interval has passed.