React to sources and git repos changing
This commit makes the syncgroup controller

 - update its sync objects when its spec changes
 - ... and when it's pointed at a GitRepository, when there's a new
   revision.

The first requires more complexity in the code for reconciling a
SyncGroup -- it will now compare each Sync that is supposed to exist
(once it's figured that which those are), and update it if it needs a
new spec.

The second requires a watch on GitRepository objects, and indexing
SyncGroups by the git repository objects they refer to.