React to clusters coming and going
When clusters arrive or depart, or change their labels, the Syncgroup
controller will want to adjust which Sync objects exist.

This commit adds a watch for Cluster objects that will find any
SyncGroups that select the cluster in question, and tell them to
reconcile. Note that on updates, the handler function is called twice,
once with the old object and once with the new -- that lets the
SyncGroup see when a cluster transitions from being selected to not
being selected.