Move site of unready cluster sync
Previously: when calculating which syncs should exist, ignore clusters
that aren't ready. But this means you don't get a sync that can have
status; so,

Now: when considering a sync, if it's cluster isn't ready, skip it.

This also changes the condition from `controlPlaneReady` to
`controlPlaneInitialized`; empirically it seems that a cluster is
ready once it's passed the latter gate, while the ready flag may not
ever be set.