Add RemoteSync, distinct to Sync
Instead of having Sync be local or target a cluster, have a separate
type for a sync that targets a cluster. Modify SyncGroup to only deal
with remote syncs (since a SyncGroup which targets the local cluster
is .. just a Sync). This changes the semantics of the sync group
selector: missing means no clusters, empty means all clusters.

This does not implement the RemoteSync controller, which will share a
lot of machinery with the Sync controller.