Draft RFC on packages
Defining dependence between syncs as in RFC 0002 is useful when you
have different repos you want to bring together. This RFC is about how
to specify dependence and so on _within_ a sync, which is more often
what you'd want to do when you're assembling a configuration from
distributable components.