Test that syncs apply to the local cluster
This adds a test to check that the sync controller will respond to a
Sync object by downloading its source and applying to the cluster
... and fixes a few things so that it passes.

One caveat is that the sync controller needs to be pointed at the test
environment, instead of assuming it's using whatever is in the
kubeconfig it finds (which is either none, or that of the user running
the test). For this purpose, the sync controller gets a
`KubernetesEndpoint` field, set to the testEnv endpoint during tests.